*replaced unnecessary Big int with long.

+++ src/main/kotlin/macarthur1337/day9/Part1.kt
@@ -2,7 +2,6 @@ package macarthur1337.day9
 
 import org.springframework.util.ResourceUtils
 import java.io.InputStream
-import java.math.BigInteger
 
 /**
  * @author N. E. Absil
@@ -11,20 +10,21 @@ import java.math.BigInteger
 fun main() {
     val inputStream: InputStream = ResourceUtils.getFile("classpath:input9").inputStream()
 
-    inputStream.bufferedReader().useLines { lines -> lines.forEach { numbers.add(it.toBigInteger()) }}
+    inputStream.bufferedReader().useLines { lines -> lines.forEach { numbers.add(it.toLong()) } }
 
     val preamble = 25
     var index = preamble
 
-    while (permutations(numbers.subList(index - preamble, index)).
-            filter { (a, b) -> a + b == numbers[index] }.any()) { index++ }
+    while (permutations(numbers.subList(index - preamble, index)).filter { (a, b) -> a + b == numbers[index] }.any()) {
+        index++
+    }
 
     println(numbers[index])
 }
 
-val numbers = mutableListOf<BigInteger>()
+val numbers = mutableListOf<Long>()
 
-fun permutations(list: MutableList<BigInteger>): Sequence<Pair<BigInteger, BigInteger>> = sequence {
+fun permutations(list: MutableList<Long>): Sequence<Pair<Long, Long>> = sequence {
     (0 until list.size).forEach { i ->
         (i + 1 until list.size).forEach { j -> yield(list[i] to list[j]) }
     }
